support two-dim input for seqlen-aware linear

+++ test/test_seqlen_dynamic_linear.py
@@ -15,6 +15,7 @@ def random_seqlen(batchsize, max_seqlen):
 
 def test_correstness(spl, seqlens, batch_size, max_seqlen, hidden_n, dtype):
     activation = torch.rand(batch_size, max_seqlen, hidden_n, dtype=dtype)
+    # activation = torch.rand(batch_size * max_seqlen, hidden_n, dtype=dtype)
     a1 = activation.clone().detach().cuda()
     a2 = activation.clone().detach().cuda()
     o1 = spl(a1)
@@ -52,7 +53,7 @@ if __name__ == '__main__':
     batch_size = 32
     max_seqlen = 128
     hidden_n = 768
-    test_type = torch.float16
+    test_type = torch.float32
     seqlens = random_seqlen(batch_size, max_seqlen).cuda()
 
     ori_linear = torch.nn.Linear(hidden_n, hidden_n, bias=True).cuda().to(test_type)
